feat: conditionally fetch workspace usage
