+++ frontend/app/project/[projectId]/layout.tsx
@@ -1,6 +1,5 @@
 import "@/app/globals.css";
 
-import { eq } from "drizzle-orm";
 import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
 import { getServerSession } from "next-auth";
@@ -13,75 +12,11 @@ import ProjectUsageBanner from "@/components/project/usage-banner";
 import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar";
 import { ProjectContextProvider } from "@/contexts/project-context";
 import { UserContextProvider } from "@/contexts/user-context";
+import { getProjectDetails } from "@/lib/actions/project";
 import { getProjectsByWorkspace } from "@/lib/actions/projects";
-import { getWorkspaceInfo, getWorkspaceUsage } from "@/lib/actions/workspace";
+import { getWorkspaceInfo } from "@/lib/actions/workspace";
 import { authOptions } from "@/lib/auth";
-import { db } from "@/lib/db/drizzle";
-import { projects, subscriptionTiers, workspaces } from "@/lib/db/migrations/schema";
 import { Feature, isFeatureEnabled } from "@/lib/features/features";
-import { GetProjectResponse } from "@/lib/workspaces/types";
-
-async function getProjectDetails(projectId: string): Promise<GetProjectResponse> {
-  const projectResult = await db
-    .select({
-      id: projects.id,
-      name: projects.name,
-      workspaceId: projects.workspaceId,
-    })
-    .from(projects)
-    .where(eq(projects.id, projectId))
-    .limit(1);
-
-  if (projectResult.length === 0) {
-    throw new Error("Project not found");
-  }
-  const project = projectResult[0];
-
-  const workspaceResult = await db
-    .select({
-      id: workspaces.id,
-      tierId: workspaces.tierId,
-    })
-    .from(workspaces)
-    .where(eq(workspaces.id, project.workspaceId))
-    .limit(1);
-
-  if (workspaceResult.length === 0) {
-    throw new Error("Workspace not found for project");
-  }
-  const workspace = workspaceResult[0];
-  const usageResult = await getWorkspaceUsage(project.workspaceId);
-
-  const tierResult = await db
-    .select({
-      name: subscriptionTiers.name,
-      stepsLimit: subscriptionTiers.steps,
-      bytesLimit: subscriptionTiers.bytesIngested,
-    })
-    .from(subscriptionTiers)
-    .where(eq(subscriptionTiers.id, workspace.tierId))
-    .limit(1);
-
-  if (tierResult.length === 0) {
-    throw new Error("Subscription tier not found for workspace");
-  }
-  const tier = tierResult[0];
-
-  // Convert bytes to GB (1 GB = 1024^3 bytes)
-  const bytesToGB = (bytes: number): number => bytes / (1024 * 1024 * 1024);
-
-  const gbUsedThisMonth = bytesToGB(usageResult.totalBytesIngested);
-  const gbLimit = bytesToGB(Number(tier.bytesLimit));
-
-  return {
-    id: project.id,
-    name: project.name,
-    workspaceId: project.workspaceId,
-    gbUsedThisMonth,
-    gbLimit,
-    isFreeTier: tier.name.toLowerCase().trim() === "free",
-  };
-}
 
 export default async function ProjectIdLayout(props: { children: ReactNode; params: Promise<{ projectId: string }> }) {
   const params = await props.params;
