+++ frontend/lib/workspaces/types.ts
@@ -41,16 +41,6 @@ export interface WorkspaceInvitation {
   createdAt: string;
 }
 
-export type GetProjectResponse = {
-  id: string;
-  name: string;
-  workspaceId: string;
-  // New GB-based usage fields
-  gbUsedThisMonth: number;
-  gbLimit: number;
-  isFreeTier: boolean;
-};
-
 export interface ProjectStats {
   datasetsCount: number;
   spansCount: number;
